Move VJPEmitter out to separate file.
