+++ lib/SILOptimizer/Utils/Differentiation/CMakeLists.txt
@@ -5,4 +5,5 @@ silopt_register_sources(
   DifferentiationInvoker.cpp
   JVPEmitter.cpp
   LinearMapInfo.cpp
+  VJPEmitter.cpp
 )
